+++ CHANGELOG.md
@@ -14,8 +14,8 @@ Changes to be including in future/planned release notes will be added here.
 - `gcp`: Dedicated `proxy_builder_sa` service account is now provisioned and used for Cloud Build operations when provisioning Cloud Functions, eliminating project-level IAM dependencies on the default Compute Engine service account.
 - `zoom`: Default rules for Zoom have been updated, removing fields and endpoints not required.
 - `pseudonymizeRegexMatches`: apply the pseudonymization to any match found in the field instead of the first one.
-- `GitLab`: added support for GitLab (gitlab.com)
 - `Atlassian Organization API`: added new connector for fetching audit events from Atlassian Organization API, which provides access to audit events from users and tools like Rovo.
+- `GitLab`: added support for GitLab (gitlab.com) and Self-Managed/Dedicated instances
 
 ## [0.5.18](https://github.com/Worklytics/psoxy/releases/tag/v0.5.18)
 - `aws`: updated IAM policy for invoking lambda functions, to reflect AWS perm changes; anyone deploying NEW api connectors on AWS will have to use this version or later, or backport similar change. (See https://docs.aws.amazon.com/lambda/latest/dg/urls-auth.html)